+++ front-end/src/api/BaseUrl.js
@@ -1,2 +1,3 @@
-export const baseUrl = 'http://localhost:8090/';
-console.log(baseUrl);
+// console.log(process.env);
+// console.log(process.env.REACT_APP_BASE_URL);
+export const baseUrl = process.env.REACT_APP_BASE_URL;
