+++ front-end/src/api/BoardApi.js
@@ -1,9 +1,9 @@
 import axios from 'axios';
 import { baseUrl } from './BaseUrl';
+// console.log(baseUrl);
 
 //board API
 export const findBoardAll = () => {
-	console.log(baseUrl);
 	const result = axios.get(baseUrl + 'board');
 	return result;
 };
