+++ front-end/src/components/Navigator.js
@@ -1,1 +1,54 @@
 //Navigator.js
+import React, { useState } from 'react';
+import {
+	MailOutlined,
+	AppstoreOutlined,
+	SettingOutlined,
+} from '@ant-design/icons';
+import Icon from '@mdi/react';
+import { Menu } from 'antd';
+
+const { SubMenu } = Menu;
+
+function Navigator() {
+	const [current, setCurrent] = useState('mail');
+	const handleClick = (e) => {
+		console.log('click ', e);
+		this.setState({ current: e.key });
+	};
+	return (
+		<Menu
+			style={{ background: 'rgb(224, 224, 232)' }}
+			onClick={handleClick}
+			selectedKeys={[current]}
+			mode="horizontal"
+		>
+			<Menu.Item key="mail" icon={<MailOutlined />}>
+				Navigation One
+			</Menu.Item>
+			<Menu.Item key="app" disabled icon={<AppstoreOutlined />}>
+				Navigation Two
+			</Menu.Item>
+			<SubMenu
+				key="SubMenu"
+				icon={<SettingOutlined />}
+				title="Navigation Three - Submenu"
+			>
+				<Menu.ItemGroup title="Item 1">
+					<Menu.Item key="setting:1">Option 1</Menu.Item>
+					<Menu.Item key="setting:2">Option 2</Menu.Item>
+				</Menu.ItemGroup>
+				<Menu.ItemGroup title="Item 2">
+					<Menu.Item key="setting:3">Option 3</Menu.Item>
+					<Menu.Item key="setting:4">Option 4</Menu.Item>
+				</Menu.ItemGroup>
+			</SubMenu>
+			<Menu.Item key="alipay">
+				<a href="https://ant.design" target="_blank" rel="noopener noreferrer">
+					Navigation Four - Link
+				</a>
+			</Menu.Item>
+		</Menu>
+	);
+}
+export default Navigator;
